del dubbo & sql

+++ asura-app/app/asura/app/api/SqlApi.scala
@@ -1,19 +1,24 @@
 package asura.app.api
 
 import akka.actor.ActorSystem
+import asura.app.AppErrorMessages
 import asura.app.api.model.TestSql
+import asura.common.model.ApiResError
 import asura.common.util.StringUtils
+import asura.core.es.EsResponse
 import asura.core.es.actor.ActivitySaveActor
 import asura.core.es.model.{Activity, SqlRequest}
-import asura.core.es.service.SqlRequestService
+import asura.core.es.service.{JobService, ScenarioService, SqlRequestService}
 import asura.core.model.QuerySqlRequest
 import asura.core.runtime.RuntimeContext
 import asura.core.sql.SqlRunner
 import asura.core.util.{JacksonSupport, JsonPathUtils}
+import asura.play.api.BaseApi.OkApiRes
 import javax.inject.{Inject, Singleton}
 import org.pac4j.play.scala.SecurityComponents
 import play.api.Configuration
-
+import asura.core.ErrorMessages
+import scala.concurrent.{ExecutionContext, Future}
 import scala.concurrent.ExecutionContext
 
 @Singleton
@@ -50,8 +55,32 @@ class SqlApi @Inject()(
     })
   }
 
-  def delete(id: String) = Action.async { implicit req =>
-    SqlRequestService.deleteDoc(id).toOkResult
+  def delete(id: String, preview: Option[Boolean]) = Action.async { implicit req =>
+    val caseIds = Seq(id)
+    val res = for {
+      s <- ScenarioService.containCase(caseIds)
+      j <- JobService.containCase(caseIds)
+    } yield (s, j)
+    res.flatMap(resTuple => {
+      val (scenarioRes, jobRes) = resTuple
+      if (scenarioRes.isSuccess && jobRes.isSuccess) {
+        if (preview.nonEmpty && preview.get) {
+          Future.successful(toActionResultFromAny(Map(
+            "scenario" -> EsResponse.toApiData(scenarioRes.result),
+            "job" -> EsResponse.toApiData(jobRes.result)
+          )))
+        } else {
+          if (scenarioRes.result.isEmpty && jobRes.result.isEmpty) {
+            SqlRequestService.deleteDoc(id).toOkResult
+          } else {
+            Future.successful(OkApiRes(ApiResError(getI18nMessage(AppErrorMessages.error_CantDeleteCase))))
+          }
+        }
+      } else {
+        val errorRes = if (!scenarioRes.isSuccess) scenarioRes else jobRes
+        ErrorMessages.error_EsRequestFail(errorRes).toFutureFail
+      }
+    })
   }
 
   def put() = Action(parse.byteString).async { implicit req =>
